# src/main/api/api.yaml
openapi: 3.0.0
info:
  version: "1.0.0"
  title: "Authorization API"
servers:
  - url: "http://localhost:8080"
    description: "Local server"
paths:
  /authorize:
    post:
      tags:
        - authorization
      summary: "Authorize a user"
      description: "Allows a user to authorize access to the API"
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                username:
                  type: string
                  description: "The username of the user"
                password:
                  type: string
                  description: "The password of the user"
              required:
                - username
                - password
      responses:
        '200':
          description: "Authorization successful"
          content:
            application/json:
              schema:
                type: object
                properties:
                  authorization_token:
                    type: string
                    description: "The authorization token"
                  role:
                    type: string
                    description: "The role of the authorized user (USER or ADMIN)"
        '403':
          description: "Authorization unsuccessful"
  /projects:
    get:
      tags:
        - projects
      summary: "get all user's projects"
      security:
        - basicAuth: [ ]
      responses:
        '200':
          description: "Project's of user retrieved successfully"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectUserCard"
        '403':
          $ref: "#/components/responses/UnauthorizedError"
    post:
      tags:
        - projects
      summary: "create new project"
      security:
        - basicAuth: [ ]
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUserCard"
      responses:
        '200':
          description: "New project created successfully"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectUserCard"
  /projects/{id}:
    get:
      tags:
        - projects
      summary: "get detailed info of project"
      security:
        - basicAuth: []
      parameters:
        - name: id
          in: path
          description: "ID of the project to retrieve"
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: "Project's info retrieved successfully"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectUserCard"
        '403':
          $ref: "#/components/responses/UnauthorizedError"
    put:
      tags:
        - projects
      summary: "edit project by user"
      security:
        - basicAuth: []
      parameters:
        - name: id
          in: path
          description: "ID of the project to edit"
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EditProjectUserCard"
      responses:
        '200':
          description: "Project's info updated successfully"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectUserCard"
        '403':
          $ref: "#/components/responses/UnauthorizedError"

    delete:
      tags:
        - projects
      summary: "delete project"
      security:
        - basicAuth: [ ]
      parameters:
        - name: id
          in: path
          description: "ID of the project to delete"
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: "Project's deleted successfully"
        '403':
          $ref: "#/components/responses/UnauthorizedError"

  /projects/{id}/rebuild:
    post:
      tags:
        - projects
      summary: "rebuild project by user"
      security:
        - basicAuth: [ ]
      parameters:
        - name: id
          in: path
          description: "ID of the project to rebuild"
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: "Project rebuild updated successfully"
        '403':
          $ref: "#/components/responses/UnauthorizedError"

  /projects/{id}/start:
    post:
      tags:
        - projects
      summary: "start project by user"
      security:
        - basicAuth: [ ]
      parameters:
        - name: id
          in: path
          description: "ID of the project to start"
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: "Project started successfully"
        '403':
          $ref: "#/components/responses/UnauthorizedError"

  /projects/{id}/stop:
    post:
      tags:
        - projects
      summary: "stop project by user"
      security:
        - basicAuth: [ ]
      parameters:
        - name: id
          in: path
          description: "ID of the project to stop"
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: "Project stopped successfully"
        '403':
          $ref: "#/components/responses/UnauthorizedError"

  /projects/{id}/logs:
    get:
      tags:
        - projects
      summary: "stop project by user"
      security:
        - basicAuth: [ ]
      parameters:
        - name: id
          in: path
          description: "ID of the project to stop"
          required: true
          schema:
            type: integer
            format: int64
        - name: startDate
          in: query
          description: "Дата с которой необходимо получить лог"
          schema:
            type: string
            format: date-time
            pattern: "yyyy-MM-dd'T'HH:mm:ss.SSSZ"
        - name: endDate
          in: query
          description: "Дата по которую необходимо получить лог"
          schema:
            type: string
            format: date-time
            pattern: "yyyy-MM-dd'T'HH:mm:ss.SSSZ"
        - name: asFile
          in: query
          description: "download logs as file. default is false"
          schema:
            type: boolean
            example: false
      responses:
        '200':
          description: "Logs retrieved successfully"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Logs"

        '403':
          $ref: "#/components/responses/UnauthorizedError"

  /projects/adminView:
    get:
      tags:
        - projects
      summary: "get all projects"
      security:
        - basicAuth: [ ]
      responses:
        '200':
          description: "Projects retrieved successfully"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminView"
        '403':
          $ref: "#/components/responses/UnauthorizedError"

  /users:
    get:
      tags:
        - users
      summary: "retrieve user's"
      security:
        - basicAuth: [ ]
      responses:
        '200':
          description: "User's retrieved successfully"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        '403':
          $ref: "#/components/responses/UnauthorizedError"
    post:
      tags:
        - users
      summary: "create new user"
      security:
        - basicAuth: [ ]
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EditUser"
      responses:
        '200':
          description: "User creation successfully"
        '403':
          $ref: "#/components/responses/UnauthorizedError"
  /users/{id}:
    put:
      tags:
        - users
      summary: "edit  user"
      security:
        - basicAuth: [ ]
      parameters:
        - name: id
          in: path
          description: "ID of user"
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EditUser"
      responses:
        '200':
          description: "User edit successfully"
        '403':
          $ref: "#/components/responses/UnauthorizedError"
    delete:
      tags:
        - users
      summary: "delete  user"
      security:
        - basicAuth: [ ]
      parameters:
        - name: id
          in: path
          description: "ID of user"
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: "User edit successfully"
        '403':
          $ref: "#/components/responses/UnauthorizedError"
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
  schemas:
    Project:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: "ID"
        name:
          type: string
          description: "Name"
        description:
          type: string
          format: email
          description: "Description"
        status:
          type: string
          enum:
            - ACTIVE
            - STOPPED
            - ERROR
          description: "Status"
    ProjectUserCard:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: "ID of project."
          example: 1
        name:
          type: string
          description: "Name"
          example: "Project name"
        description:
          type: string
          format: email
          description: "Description"
          example: "Description"
        status:
          type: string
          enum:
            - ACTIVE
            - STOPPED
            - ERROR
          description: "Status"
          example: "ACTIVE"
        type:
          type: string
          enum:
            - DOCKER
            - PYTHON
            - JAVA
          example: "JAVA"
        gitUrl:
          type: string
          example: "https://github.com/bratzelk/spring-boot-hello-world.git"
          description: "git repository of deployed app"
        gitBranch:
          type: string
          example: "master"
          description: "git branch of deployed app"
        ownerId:
          type: number
          example: 1
          description: "user id of project's owner"
        ownerName:
          type: string
          example: "user1"
          description: "username of project's owner"
        ports:
          type: array
          items:
            type: string
            format: int32
            example: "8080:8080"
        envs:
          type: array
          items:
            type: string
            example: "SERVER_PORT=8080"
        createdDate:
          type: string
          format: date-time
          pattern: "yyyy-MM-dd'T'HH:mm:ss.SSSZ"
          example: "2023-05-01T14:30:00.000-0700"
        cpuConsumption:
          type: number
          format: double
          example: 0.5
          description: "cpu consumption metric at this time"
        cpuRequest:
          type: number
          format: double
          example: 1
          description: "cpu requested for this project"
        ramConsumption:
          type: number
          format: double
          example: 0.5
          description: "ram consumption metric at this time"
        ramRequest:
          type: number
          format: double
          example: 1
          description: "ram requested for this project"
        storageConsumption:
          type: number
          format: double
          example: 0.5
          description: "storage consumption metric at this time"
        storageRequest:
          type: number
          format: double
          example: 1
          description: "storage requested for this project"
    EditProjectUserCard:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: "ID of project."
          example: 1
        name:
          type: string
          description: "Name"
          example: "Project name"
        description:
          type: string
          format: email
          description: "Description"
          example: "Description"
        type:
          type: string
          enum:
            - DOCKER
            - PYTHON
            - JAVA
          example: "JAVA"
        gitUrl:
          type: string
          example: "https://github.com/bratzelk/spring-boot-hello-world.git"
          description: "git repository of deployed app"
        gitBranch:
          type: string
          example: "master"
          description: "git branch of deployed app"
        ports:
          type: array
          items:
            type: string
            format: int32
            example: "8080:8080"
        envs:
          type: array
          items:
            type: string
            example: "SERVER_PORT=8080"
    AdminView:
      type: object
      properties:
        clusterStatistics:
          type: object
          properties:
            currentCpu:
              type: number
              example: 4
            availibleCpu:
              type: number
              example: 10
            currentRam:
              type: number
              example: 10
            availibleRam:
              type: number
              example: 20
            currentStorage:
              type: number
              example: 20
            availibleStorage:
              type: number
              example: 100
        projects:
          type: array
          items:
            $ref: "#/components/schemas/ProjectUserCard"
    Logs:
      type: object
      properties:
        value:
          type: string
          example: "2019-03-05 10:57:51.112  INFO 45469 --- [           main] org.apache.catalina.core.StandardEngine  : Starting Servlet Engine: Apache Tomcat/7.0.52 \n 2019-03-05 10:57:51.253  INFO 45469 --- [ost-startStop-1] o.a.c.c.C.[Tomcat].[localhost].[/]       : Initializing Spring embedded WebApplicationContext \n 2019-03-05 10:57:51.253  INFO 45469 --- [ost-startStop-1] o.s.web.context.ContextLoader            : Root WebApplicationContext: initialization completed in 1358 ms \n 2019-03-05 10:57:51.698  INFO 45469 --- [ost-startStop-1] o.s.b.c.e.ServletRegistrationBean        : Mapping servlet: 'dispatcherServlet' to [/] \n 2019-03-05 10:57:51.702  INFO 45469 --- [ost-startStop-1] o.s.b.c.embedded.FilterRegistrationBean  : Mapping filter: 'hiddenHttpMethodFilter' to: [/*] UnauthorizedError: description: Unauthorized access error"
    User:
      type: object
      properties:
        id:
          type: number
          example: 1
        name:
          type: string
          example: "user1"
        userRole:
          type: string
          enum:
            - USER
            - ADMIN
        createdDate:
          type: string
          format: date-time
          pattern: "yyyy-MM-dd'T'HH:mm:ss.SSSZ"
          example: "2023-05-01T14:30:00.000-0700"
        currentCpu:
          type: number
          example: 4
        requestCpu:
          type: number
          example: 8
        availibleCpu:
          type: number
          example: 10
        currentRam:
          type: number
          example: 10
        requestRam:
          type: number
          example: 15
        availibleRam:
          type: number
          example: 20
        currentStorage:
          type: number
          example: 20
        requestStorage:
          type: number
          example: 50
        availibleStorage:
          type: number
          example: 100
    EditUser:
      type: object
      properties:
        name:
          type: string
          example: "user1"
        password:
          type: string
          example: "qwe123"
        currentCpu:
          type: number
          example: 4
        requestCpu:
          type: number
          example: 8
        availibleCpu:
          type: number
          example: 10
        currentRam:
          type: number
          example: 10
        requestRam:
          type: number
          example: 15
        availibleRam:
          type: number
          example: 20
        currentStorage:
          type: number
          example: 20
        requestStorage:
          type: number
          example: 50
        availibleStorage:
          type: number
          example: 100


  responses:
    UnauthorizedError:
      description: "Unauthorized access error"
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                description: "The error message"
          example:
            error: "Unauthorized: Invalid Auth token"
      
    